backend/requirements.txt
# Core framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic[email]
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# AI/ML
numpy==1.26.4

# Distributed system layer
redis==5.0.1
aio-pika==9.4.0
tenacity==8.2.3
structlog==24.1.0